fix(keirin): 三連複payout計算を10円単位切り捨てに修正
odds_value*100をround()するだけだと浮動小数点誤差や
Winticketオッズの精度により1の位が0にならない払戻額が
生成され、年間サマリーの合計が10円単位からずれる不具合が
あった。公式払戻ルール通り10円単位への切り捨てに修正。

+++ backend/src/api/keirin_router.py
@@ -496,7 +496,8 @@ async def refresh_picks(
         for odds_row in odds_rows:
             try:
                 nums = [int(p) for p in _re.split(r"[-=→]", str(odds_row["combination"])) if p]
-                trio_map[frozenset(nums)] = int(round(float(odds_row["odds_value"]) * 100))
+                # 公式払戻金は10円単位に切り捨て。round()で浮動小数点誤差を吸収してから10円に丸める
+                trio_map[frozenset(nums)] = round(float(odds_row["odds_value"]) * 100) // 10 * 10
             except (ValueError, TypeError):
                 continue
